MarinaMaster-Multi-Location-Microsoft-DynamicImageSize/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.4",
    "pillow>=10.4.0",
    "pillow-heif>=0.16.0",
    "requests>=2.32.4",
    "vision-agent>=1.1.17",
]
